(#47) - Add support _db_updates feeds
Follow detects whether an url (or db or uri) ends with `/_db_updates` and changes
its internals accordingly.

There is no need for extending the API and the change is fully backwards compatibel,
because the can't exist a db called `_db_updates`.
With one exception: If a user was to route a couchdb database on an endpoint
which is `_db_updates` it will not work the same anymore :)